// client/src/utils/enums.ts
export enum Mode {
	Connect = 'connect',
	Move = 'move',
	Add = 'add',
	StartSelect = 'start-select',
	DestinationSelect = 'destination-select',
}

export enum Terminal {
    Start = 'start',
    Destination = 'destination',
}

export enum EdgeType {
	Weighted = 'weighted',
	Unweighted = 'unweighted',
}